Move domain ability traits into traits namespace
